requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asyncmy==0.2.10
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.116.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
passlib==1.7.4
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
setuptools==78.1.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
wheel==0.45.1
